Split traits Authenticates from Registers and add abstract methods: validator and create
